Finished adding the module documentation for eigrp_classic. I also fixed an oversight related to the output of module.run_command.
It gives three outputs, one for return code with zero indicating no error, stdout indicating what I actually wanted, and stderr that returns the error message in the event of an error.

Lastly I added an ansible integration test with an assertion that needs t be correct for the test to pass.

TODO: The rest of the collection documentation is still default.
In order to make a release, I need to implement a github workflow